% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0008_ontology.R
\name{read_acm0008_technology_shapes}
\alias{read_acm0008_technology_shapes}
\title{Load ACM0008 technology applicability shape}
\usage{
read_acm0008_technology_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shaclR:read_shacl]{shaclR::read_shacl()}}.
}
\description{
Load ACM0008 technology applicability shape
}
